remove old unused and erroring avatar select menu
